requirements.txt
pandas
yfinance
matplotlib
PyYAML
statsmodels
numpy 